# Conflicts:
#	.gitignore

+++ core/src/view/View3D.cpp
@@ -650,7 +650,7 @@ void view::View3D::SetCursor2DPosition(float x, float y) {
         // stuff from protein::View3DMouse
         CallRender3D *cr3d = this->rendererSlot.CallAs<CallRender3D>();
         if (cr3d) {
-            cr3d->SetMouseInfo(static_cast<int>(x), static_cast<int>(y),
+            cr3d->SetMouseInfo(static_cast<float>(static_cast<int>(x)), static_cast<float>(static_cast<int>(y)),
                 this->mouseFlags);
             if ((*cr3d)(3)) {
                 this->mouseX = (float)static_cast<int>(x);
